refactor(selfhost): extract Check/Resolve types to selfhost/api
Factor the pure-data check and resolve return shapes into a new
`internal/selfhost/api` subpackage, then re-export them from
`internal/selfhost` via type aliases. Existing callers
(`selfhost.CheckResult` etc.) continue to compile unchanged; callers
that want a boundary-only dependency can now import
`internal/selfhost/api` directly without dragging in the generated
bootstrap core.

Scope:
- Moved (no methods, self-contained): CheckSummary, CheckedNode,
  CheckedBinding, CheckedSymbol, CheckInstantiation,
  CheckDiagnosticRecord, CheckResult, ResolveSummary, ResolvedSymbol,
  ResolvedRef, ResolvedTypeRef, ResolveDiagnosticRecord, ResolveResult.
- Left in place: PackageCheck* / PackageResolve* / CfgEnv — these
  entangle with `*ast.File`, `sourcemap.Map`, and selfhost-internal
  receiver methods. Follow-up PR can split them out after teasing apart
  the method dependencies.
